New script remaned to old, to suggest modification
