+++ README.md
@@ -6,24 +6,13 @@ Six arguments need to be specified for running the script: your valid email addr
 
 If you dont have an NCBI api key, create an account at https://www.ncbi.nlm.nih.gov/account/ using your email. Once you sign in, go to the top right corner where your email id displayed. Click on it and you will be directed to a new page where you can find your api key. Copy the key to use it as an argument for the script. 
 
-Usage: python fetch_contigs.py "Valid email id" "NCBI api_key" "Genus name" "Species Name" "Number of assemblies" 
+Usage: python fetch_contigs.py "Valid email id" "NCBI api_key" "Genus name" "Species Name" "Number of assemblies" "Output directory"  
+Output/s: zipped fasta assembly files inside the output directory
 
+**Exercise 2**: Python script for analyzing fasta assemblies and calculating N50 lengths, summary statistics and plotting their distribution  
 
-**Exercise 2**: Python script for analyzing fasta assemblies and calculating N50 contig length
-
-Given a directory of fasta assembly files, the script reads each assembly file and embedded fasta formatted contigs to calculate N50 values. 
+Given a directory of fasta assembly files, the script reads each assembly file and embedded fasta formatted contigs to calculate N50 values. Subsequently, calculates summary statistics using *pandas.Series.Describe* function and generates a histogram of N50 assembly lengths using *matplotlib* 
 
 Usage: python calculating_n50_assemblies.py INPUT_DIRECTORY OUTPUT_DIRECTORY  
+Output/s: *summary_statistics.txt* file containing summary statistics and *hist.pdf* plot showing histogram   
 
-The program needs two arguments corresponding to input directory containing fasta assembly files and output directory where output file "N50.txt" is written with assembly name and N50 lengths in tab separated value format:   
-   
-RSZF02  91963   
-RTBZ02  71579   
-RTFI02  93615   
-..............
-..............
-
-**Exercise 3**: Python script for calculating summary statistics of N50 and plotting its distribution   
-Extends exercise 2 by incorporating a function to calculate summary statistics using *pandas.Series.Describe* function. Also generates a histogram of N50 assembly lengths using *matplotlib*
-
-Usage: python n50_statistics.py INPUT_DIRECTORY OUTPUT_DIRECTORY
